Add docker-compose file for integration tests
