+++ .travis.yml
@@ -27,5 +27,6 @@ before_script:
 script:
   - npm run test
   - npm run test:deployment_tool
+  - npm run test:integration
 after_script:
   - kill $(ps aux | grep 'ganache-cli' | awk '{print $2}')
