+++ test_integration/docker-compose.yml
@@ -1,0 +1,13 @@
+version: "3"
+
+services:
+
+  geth_node_origin:
+    image: augurproject/dev-node-geth:v1.8.18
+    ports:
+      - "8545:8545"
+
+  geth_node_auxiliary:
+    image: augurproject/dev-node-geth:v1.8.18
+    ports:
+      - "8546:8545"
